requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
decorator==4.4.2
future==0.18.2
idna==2.9
pycparser==2.20
python-telegram-bot==12.5.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.8
